# stario-platform/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Stario API
  description: |
    Stario / TabrikStar - AI Emotion Platform API

    Create personalized AI video greetings, face quizzes, posters, and more
    with your favorite artists.

    ## Authentication
    Most endpoints require Bearer token authentication.
    Get your token via `/auth/login` or `/auth/telegram`.

    ## Rate Limiting
    - 60 requests per minute per user
    - 10,000 requests per day per user

    ## Compliance (Uzbekistan 2025-2030)
    - Data residency: All user data stored within Uzbekistan
    - PII minimization: Face Quiz photos deleted within 3 seconds
    - Audit logs retained for 90 days

  version: 1.0.0
  contact:
    name: Stario Support
    email: support@stario.uz
servers:
  - url: https://api.stario.uz
    description: Production
  - url: https://api-staging.stario.uz
    description: Staging
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /videos/generate:
    post:
      tags: [Videos]
      summary: Start video generation
      description: |
        Generate a personalized "Stario Moment" video greeting.
        Target processing time: ≤40 seconds.
      operationId: generateVideo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoGenerationRequest'
      responses:
        '200':
          description: Generation job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoGenerationResponse'
  /videos/jobs/{jobId}:
    get:
      tags: [Videos]
      summary: Get video job status
      operationId: getVideoJobStatus
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoJobStatus'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VideoGenerationRequest:
      type: object
      required: [artist_id, custom_message, recipient_name]
      properties:
        artist_id:
          type: string
        prompt_template_id:
          type: string
        custom_message:
          type: string
          maxLength: 500
        recipient_name:
          type: string
        occasion:
          type: string
          default: greeting
        language:
          type: string
          default: uz
        duration_seconds:
          type: integer
          enum: [15, 30, 60]
          default: 15
    VideoGenerationResponse:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
        estimated_duration_seconds:
          type: integer
        queue_position:
          type: integer
    VideoJobStatus:
      type: object
      properties:
        job_id:
          type: string
        status:
          type: string
          enum: [pending, queued, processing, completed, failed]
        progress_percent:
          type: integer
        video_url:
          type: string
        thumbnail_url:
          type: string
        error_message:
          type: string
...